requirements.txt
flask
flask-cors
google-auth
google-auth-oauthlib
google-api-python-client
requests-oauthlib
oauthlib
sqlalchemy
psycopg2-binary
redis
bcrypt